refactor: parentTagIcon
