+++ libs/journeys/ui/src/components/ParentTagIcon/ParentTagIcon.spec.tsx
